+++ lib/app_module.dart
@@ -1,0 +1,3 @@
+import 'package:flutter_modular/flutter_modular.dart';
+
+class AppModule extends Module {}
