+++ pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   path: ^1.8.0
   path_provider: ^2.0.7
   equatable: ^2.0.3
+  flutter_modular: ^4.2.0
 
 dev_dependencies:
   flutter_test:
